mand done

+++ srcs/died.c
@@ -11,7 +11,7 @@ int64_t	is_dead(t_data *data)
 		+ mtx_do_func(&data->phi->end_of_simulation, &data->phi->eos, ret_arg));
 }
 
-int64_t	ret_arg(int32_t *arg)
+int64_t	ret_arg(int64_t *arg)
 {
 	return (*arg);
 }
@@ -22,7 +22,7 @@ void	i_have_died(t_data *data)
 	mtx_do_func(&data->phi->end_of_simulation, &data->phi->eos, set_flag);
 }
 
-int64_t	set_flag(int32_t *flag)
+int64_t	set_flag(int64_t *flag)
 {
 	*flag = 1;
 	return (0);
